Added configurable messages
Messages can be configured in 'messages.lang'.
By default it is only stored in the jar. It is recommended
that you simply create a 'messages.lang' file in the data folder
and only add the parts you wish to change. Any messages not present
will be filled in with the default value